+++ src/main/java/org/manifold/compiler/middle/serialization/SchematicSerializer.java
@@ -4,7 +4,7 @@ import org.manifold.compiler.middle.Schematic;
 
 import com.google.gson.JsonObject;
 
-public class SchematicSerializer implements SeralizationConsts {
+public class SchematicSerializer implements SerializationConsts {
 
   public JsonObject serialize(Schematic sch) {
     JsonObject schJson = new JsonObject();
